Update main.cpp
Created function to print board, dynamically allocate space for 2D array in order to pass the board pointer to functions. Added skeleton for check_win function which returns a boolean and takes the board as a argument suggest first checking for the easy wins i.e. horizontal and vertical and then check for diagonal. Commented out system call as I haven't looked into getting it to work on VS.